drivers: imx: esai: Fix watermark
My understanding of the watermark was backwards. For TX, the watermark
triggers when you have TFWM slots empty, while for RX the watermark
triggers when you have RFWM slots full. In both cases it's the number of
slots you need to transfer. My old understanding was that it was the
number of full slots in both cases.

Signed-off-by: Paul Olaru <paul.olaru@nxp.com>