[DO NOT MERGE] Disable volume component, convert it to passthrough
This seems to lower how much pipeline_copy() takes enough to remove all
issues of missed IPCs. It is required on this build if built with gcc
(and thus using volume_generic) as with this configuration it seems that
the volume processing takes up enough time to make the DSP fail to
always send the host position IPC properly, which induces weird playback
issues (host buffer repeating 2+ times every single time).

The real solution isn't this, of course. I believe something is broken
in the scheduler implementation.

Signed-off-by: Paul Olaru <paul.olaru@nxp.com>