[THIS BRANCH IS NO LONGER TRULY NEEDED]
Playback using the ESAI with the cs42888 codec works with
upstream/master and our kernel.

I think I still have a few things about DMA trace in here but for the
most part you may now ignore this branch.

This commit was originally an empty commit, I need to clean up this
branch to see if there is anything useful (DMA trace and serial console)
on this branch.

Signed-off-by: Paul Olaru <paul.olaru@nxp.com.