[DO NOT MERGE] [WIP] [HACK] platform: imx: Panic support
This commit brings some changes to the trace system as well as a hacky
way to support DSP panics. There is also an associated Linux patch,
however this commit may simply be discarded when a better way to report
DSP panics is found.

Also have the mailbox trace mechanism work side-by-side with
__dsp_printf.

Signed-off-by: Paul Olaru <paul.olaru@nxp.com>